+++ src/app/website/documentation/home/HomeView.ts
@@ -1,0 +1,35 @@
+import { Component, ElementRef } from "@angular/core";
+import { CORE_DIRECTIVES } from "@angular/common";
+import { Title } from "@angular/platform-browser";
+import { RouteConfig, RouterOutlet, RouterLink } from "@angular/router-deprecated";
+
+import $ from "jquery";
+import "semantic-ui/semantic";
+
+import template from "./template.html!";
+import "./style.css!";
+
+@Component( {
+	selector: "documents-list",
+	template: template,
+	directives: [ CORE_DIRECTIVES, RouterLink ],
+	providers: [ Title ],
+} )
+export default class HomeView {
+	element:ElementRef;
+	$element:JQuery;
+	title:Title;
+
+	constructor( element:ElementRef, title:Title ) {
+		this.element = element;
+		this.title = title;
+	}
+
+	ngAfterViewInit():void {
+		this.$element = $( this.element.nativeElement );
+	}
+
+	routerOnActivate():void {
+		this.title.setTitle( "Documentation" );
+	}
+}