+++ src/app/website/home/HomeView.ts
@@ -44,7 +44,7 @@ export default class HomeView {
 	}
 
 	routerOnActivate():void {
-		this.title.setTitle( "Home" );
+		this.title.setTitle( "CarbonLDP | Home" );
 	}
 
 	routerOnDeactivate():void {